Enum. And returning after a long winter

+++ fixtures/func_call/func_call.cpp
@@ -22,11 +22,11 @@ class ClassTest
 int main() 
 {
 	int x;
-	int a = 3;
+	int a = (int) 3.0;
 	int b = 5;
 	int c = soma(a, b);
 	
 	int d = soma(a, b, 0.3);
 	
-	int e = ClassTest::staticFunc(8);
+	int e = ClassTest::staticFunc( (int) 8.0);
 }
